manager: Fix UART connection issue
Delete the UART file descriptor object, so that the agent OS can safely
remove it from the system (this happens for some boards, like coraz7s,
when power cycling). After power cycling the board, the UART is
reconnected to the same file descriptor (and the object is reinitialized
in manager.py).

Signed-off-by: bia1708 <bianca-maria.popu@analog.com>